Full Fit, Normalized Data, Averaged Over, Bug Fixes
So this is as far as I got in fixing all the bugs that existed in the
code, as well as normalizing the data at the place I believe we need to
normalize it at. I also made it run the full fit by simply setting the
train percentage to 100%. Not that this makes it impossible to calculate
the R^2 because the model cannot be created with the function, so I made
some workarounds inside the code.

Next steps: Move to N number of inputs, will require some serious
readjustment so I will likely take forever.