Fixed Normalization, Now Works with 2 Inputs
I also changed the R^2 calculation to the Adjusted value, commented out
the R^2 for the original function since it cannot be calculated using a
linear model, added a function that makes the output cell into a matrix
with which Matlab can actually work (Line 61), removed uncessary
variables for outputs.

Now I will need to work on reading in data and making sure that the
input matrix can be expanded to any number given.